rebase

+++ walletunlocker/service.go
@@ -1,6 +1,7 @@
 package walletunlocker
 
 import (
+	"context"
 	"crypto/rand"
 	"errors"
 	"fmt"
@@ -10,12 +11,32 @@ import (
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcwallet/wallet"
 	"github.com/lightningnetwork/lnd/aezeed"
+	"github.com/lightningnetwork/lnd/chanbackup"
 	"github.com/lightningnetwork/lnd/keychain"
 	"github.com/lightningnetwork/lnd/lnrpc"
 	"github.com/lightningnetwork/lnd/lnwallet"
 	"github.com/lightningnetwork/lnd/lnwallet/btcwallet"
-	"golang.org/x/net/context"
+	"github.com/lightningnetwork/lnd/macaroons"
 )
+
+var (
+	// ErrUnlockTimeout signals that we did not get the expected unlock
+	// message before the timeout occurred.
+	ErrUnlockTimeout = errors.New("got no unlock message before timeout")
+)
+
+// ChannelsToRecover wraps any set of packed (serialized+encrypted) channel
+// back ups together. These can be passed in when unlocking the wallet, or
+// creating a new wallet for the first time with an existing seed.
+type ChannelsToRecover struct {
+	// PackedMultiChanBackup is an encrypted and serialized multi-channel
+	// backup.
+	PackedMultiChanBackup chanbackup.PackedMulti
+
+	// PackedSingleChanBackups is a series of encrypted and serialized
+	// single-channel backup for one or more channels.
+	PackedSingleChanBackups chanbackup.PackedSingles
+}
 
 // WalletInitMsg is a message sent by the UnlockerService when a user wishes to
 // set up the internal wallet for the first time. The user MUST provide a
@@ -36,6 +57,15 @@ type WalletInitMsg struct {
 	// recovery should be attempted, such as after the wallet's initial
 	// creation.
 	RecoveryWindow uint32
+
+	// ChanBackups a set of static channel backups that should be received
+	// after the wallet has been initialized.
+	ChanBackups ChannelsToRecover
+
+	// StatelessInit signals that the user requested the daemon to be
+	// initialized stateless, which means no unencrypted macaroons should be
+	// written to disk.
+	StatelessInit bool
 }
 
 // WalletUnlockMsg is a message sent by the UnlockerService when a user wishes
@@ -53,12 +83,25 @@ type WalletUnlockMsg struct {
 	// creation, but before any addresses have been created.
 	RecoveryWindow uint32
 
-	// Wallet is the loaded and unlocked Wallet. This is returned
-	// through the channel to avoid it being unlocked twice (once to check
-	// if the password is correct, here in the WalletUnlocker and again
-	// later when lnd actually uses it). Because unlocking involves scrypt
-	// which is resource intensive, we want to avoid doing it twice.
+	// Wallet is the loaded and unlocked Wallet. This is returned through
+	// the channel to avoid it being unlocked twice (once to check if the
+	// password is correct, here in the WalletUnlocker and again later when
+	// lnd actually uses it). Because unlocking involves scrypt which is
+	// resource intensive, we want to avoid doing it twice.
 	Wallet *wallet.Wallet
+
+	// ChanBackups a set of static channel backups that should be received
+	// after the wallet has been unlocked.
+	ChanBackups ChannelsToRecover
+
+	// UnloadWallet is a function for unloading the wallet, which should
+	// be called on shutdown.
+	UnloadWallet func() error
+
+	// StatelessInit signals that the user requested the daemon to be
+	// initialized stateless, which means no unencrypted macaroons should be
+	// written to disk.
+	StatelessInit bool
 }
 
 // UnlockerService implements the WalletUnlocker service used to provide lnd
@@ -74,21 +117,34 @@ type UnlockerService struct {
 	// sent.
 	UnlockMsgs chan *WalletUnlockMsg
 
-	chainDir      string
-	netParams     *chaincfg.Params
+	// MacResponseChan is the channel for sending back the admin macaroon to
+	// the WalletUnlocker service.
+	MacResponseChan chan []byte
+
+	chainDir       string
+	noFreelistSync bool
+	netParams      *chaincfg.Params
+
+	// macaroonFiles is the path to the three generated macaroons with
+	// different access permissions. These might not exist in a stateless
+	// initialization of lnd.
 	macaroonFiles []string
 }
 
 // New creates and returns a new UnlockerService.
-func New(chainDir string, params *chaincfg.Params,
+func New(chainDir string, params *chaincfg.Params, noFreelistSync bool,
 	macaroonFiles []string) *UnlockerService {
 
 	return &UnlockerService{
-		InitMsgs:      make(chan *WalletInitMsg, 1),
-		UnlockMsgs:    make(chan *WalletUnlockMsg, 1),
-		chainDir:      chainDir,
-		netParams:     params,
-		macaroonFiles: macaroonFiles,
+		InitMsgs:   make(chan *WalletInitMsg, 1),
+		UnlockMsgs: make(chan *WalletUnlockMsg, 1),
+
+		// Make sure we buffer the channel is buffered so the main lnd
+		// goroutine isn't blocking on writing to it.
+		MacResponseChan: make(chan []byte, 1),
+		chainDir:        chainDir,
+		netParams:       params,
+		macaroonFiles:   macaroonFiles,
 	}
 }
 
@@ -100,13 +156,13 @@ func New(chainDir string, params *chaincfg.Params,
 // Once the cipherseed is obtained and verified by the user, the InitWallet
 // method should be used to commit the newly generated seed, and create the
 // wallet.
-func (u *UnlockerService) GenSeed(ctx context.Context,
+func (u *UnlockerService) GenSeed(_ context.Context,
 	in *lnrpc.GenSeedRequest) (*lnrpc.GenSeedResponse, error) {
 
 	// Before we start, we'll ensure that the wallet hasn't already created
 	// so we don't show a *new* seed to the user if one already exists.
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, 0)
+	loader := wallet.NewLoader(u.netParams, netDir, u.noFreelistSync, 0)
 	walletExists, err := loader.WalletExists()
 	if err != nil {
 		return nil, err
@@ -167,6 +223,43 @@ func (u *UnlockerService) GenSeed(ctx context.Context,
 	}, nil
 }
 
+// extractChanBackups is a helper function that extracts the set of channel
+// backups from the proto into a format that we'll pass to higher level
+// sub-systems.
+func extractChanBackups(chanBackups *lnrpc.ChanBackupSnapshot) *ChannelsToRecover {
+	// If there aren't any populated channel backups, then we can exit
+	// early as there's nothing to extract.
+	if chanBackups == nil || (chanBackups.SingleChanBackups == nil &&
+		chanBackups.MultiChanBackup == nil) {
+		return nil
+	}
+
+	// Now that we know there's at least a single back up populated, we'll
+	// extract the multi-chan backup (if it's there).
+	var backups ChannelsToRecover
+	if chanBackups.MultiChanBackup != nil {
+		multiBackup := chanBackups.MultiChanBackup
+		backups.PackedMultiChanBackup = chanbackup.PackedMulti(
+			multiBackup.MultiChanBackup,
+		)
+	}
+
+	if chanBackups.SingleChanBackups == nil {
+		return &backups
+	}
+
+	// Finally, we can extract all the single chan backups as well.
+	for _, backup := range chanBackups.SingleChanBackups.ChanBackups {
+		singleChanBackup := backup.ChanBackup
+
+		backups.PackedSingleChanBackups = append(
+			backups.PackedSingleChanBackups, singleChanBackup,
+		)
+	}
+
+	return &backups
+}
+
 // InitWallet is used when lnd is starting up for the first time to fully
 // initialize the daemon and its internal wallet. At the very least a wallet
 // password must be provided. This will be used to encrypt sensitive material
@@ -184,7 +277,7 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 
 	// Make sure the password meets our constraints.
 	password := in.WalletPassword
-	if err := validatePassword(password); err != nil {
+	if err := ValidatePassword(password); err != nil {
 		return nil, err
 	}
 
@@ -198,7 +291,9 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 	// We'll then open up the directory that will be used to store the
 	// wallet's files so we can check if the wallet already exists.
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, uint32(recoveryWindow))
+	loader := wallet.NewLoader(
+		u.netParams, netDir, u.noFreelistSync, uint32(recoveryWindow),
+	)
 
 	walletExists, err := loader.WalletExists()
 	if err != nil {
@@ -231,11 +326,35 @@ func (u *UnlockerService) InitWallet(ctx context.Context,
 		Passphrase:     password,
 		WalletSeed:     cipherSeed,
 		RecoveryWindow: uint32(recoveryWindow),
+		StatelessInit:  in.StatelessInit,
 	}
 
-	u.InitMsgs <- initMsg
+	// Before we return the unlock payload, we'll check if we can extract
+	// any channel backups to pass up to the higher level sub-system.
+	chansToRestore := extractChanBackups(in.ChannelBackups)
+	if chansToRestore != nil {
+		initMsg.ChanBackups = *chansToRestore
+	}
 
-	return &lnrpc.InitWalletResponse{}, nil
+	// Deliver the initialization message back to the main daemon.
+	select {
+	case u.InitMsgs <- initMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work and to get the admin macaroon. Once the response
+		// arrives, we directly forward it to the client.
+		select {
+		case adminMac := <-u.MacResponseChan:
+			return &lnrpc.InitWalletResponse{
+				AdminMacaroon: adminMac,
+			}, nil
+
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
 // UnlockWallet sends the password provided by the incoming UnlockWalletRequest
@@ -248,7 +367,9 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 	recoveryWindow := uint32(in.RecoveryWindow)
 
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, recoveryWindow)
+	loader := wallet.NewLoader(
+		u.netParams, netDir, u.noFreelistSync, recoveryWindow,
+	)
 
 	// Check if wallet already exists.
 	walletExists, err := loader.WalletExists()
@@ -275,14 +396,36 @@ func (u *UnlockerService) UnlockWallet(ctx context.Context,
 		Passphrase:     password,
 		RecoveryWindow: recoveryWindow,
 		Wallet:         unlockedWallet,
+		UnloadWallet:   loader.UnloadWallet,
+		StatelessInit:  in.StatelessInit,
 	}
 
-	// At this point we was able to open the existing wallet with the
+	// Before we return the unlock payload, we'll check if we can extract
+	// any channel backups to pass up to the higher level sub-system.
+	chansToRestore := extractChanBackups(in.ChannelBackups)
+	if chansToRestore != nil {
+		walletUnlockMsg.ChanBackups = *chansToRestore
+	}
+
+	// At this point we were able to open the existing wallet with the
 	// provided password. We send the password over the UnlockMsgs
 	// channel, such that it can be used by lnd to open the wallet.
-	u.UnlockMsgs <- walletUnlockMsg
+	select {
+	case u.UnlockMsgs <- walletUnlockMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work. But we don't need the returned macaroon for this
+		// operation, so we read it but then discard it.
+		select {
+		case <-u.MacResponseChan:
+			return &lnrpc.UnlockWalletResponse{}, nil
 
-	return &lnrpc.UnlockWalletResponse{}, nil
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
 // ChangePassword changes the password of the wallet and sends the new password
@@ -292,7 +435,7 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 	in *lnrpc.ChangePasswordRequest) (*lnrpc.ChangePasswordResponse, error) {
 
 	netDir := btcwallet.NetworkDir(u.chainDir, u.netParams)
-	loader := wallet.NewLoader(u.netParams, netDir, 0)
+	loader := wallet.NewLoader(u.netParams, netDir, u.noFreelistSync, 0)
 
 	// First, we'll make sure the wallet exists for the specific chain and
 	// network.
@@ -316,7 +459,7 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 	}
 
 	// Make sure the new password meets our constraints.
-	if err := validatePassword(in.NewPassword); err != nil {
+	if err := ValidatePassword(in.NewPassword); err != nil {
 		return nil, err
 	}
 
@@ -325,18 +468,32 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 	if err != nil {
 		return nil, err
 	}
-	// Unload the wallet to allow lnd to open it later on.
-	defer loader.UnloadWallet()
 
-	// Since the macaroon database is also encrypted with the wallet's
-	// password, we'll remove all of the macaroon files so that they're
-	// re-generated at startup using the new password. We'll make sure to do
-	// this after unlocking the wallet to ensure macaroon files don't get
-	// deleted with incorrect password attempts.
-	for _, file := range u.macaroonFiles {
-		err := os.Remove(file)
-		if err != nil && !os.IsNotExist(err) {
-			return nil, err
+	// Now that we've opened the wallet, we need to close it in case of an
+	// error. But not if we succeed, then the caller must close it.
+	orderlyReturn := false
+	defer func() {
+		if !orderlyReturn {
+			_ = loader.UnloadWallet()
+		}
+	}()
+
+	// Before we actually change the password, we need to check if all flags
+	// were set correctly. The content of the previously generated macaroon
+	// files will become invalid after we generate a new root key. So we try
+	// to delete them here and they will be recreated during normal startup
+	// later. If they are missing, this is only an error if the
+	// stateless_init flag was not set.
+	if in.NewMacaroonRootKey || in.StatelessInit {
+		for _, file := range u.macaroonFiles {
+			err := os.Remove(file)
+			if err != nil && !in.StatelessInit {
+				return nil, fmt.Errorf("could not remove "+
+					"macaroon file: %v. if the wallet "+
+					"was initialized stateless please "+
+					"add the --stateless_init "+
+					"flag", err)
+			}
 		}
 	}
 
@@ -351,15 +508,90 @@ func (u *UnlockerService) ChangePassword(ctx context.Context,
 			"%v", err)
 	}
 
+	// The next step is to load the macaroon database, change the password
+	// then close it again.
+	// Attempt to open the macaroon DB, unlock it and then change
+	// the passphrase.
+	macaroonService, err := macaroons.NewService(
+		netDir, "lnd", in.StatelessInit,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	err = macaroonService.CreateUnlock(&privatePw)
+	if err != nil {
+		closeErr := macaroonService.Close()
+		if closeErr != nil {
+			return nil, fmt.Errorf("could not create unlock: %v "+
+				"--> follow-up error when closing: %v", err,
+				closeErr)
+		}
+		return nil, err
+	}
+	err = macaroonService.ChangePassword(privatePw, in.NewPassword)
+	if err != nil {
+		closeErr := macaroonService.Close()
+		if closeErr != nil {
+			return nil, fmt.Errorf("could not change password: %v "+
+				"--> follow-up error when closing: %v", err,
+				closeErr)
+		}
+		return nil, err
+	}
+
+	// If requested by the user, attempt to replace the existing
+	// macaroon root key with a new one.
+	if in.NewMacaroonRootKey {
+		err = macaroonService.GenerateNewRootKey()
+		if err != nil {
+			closeErr := macaroonService.Close()
+			if closeErr != nil {
+				return nil, fmt.Errorf("could not generate "+
+					"new root key: %v --> follow-up error "+
+					"when closing: %v", err, closeErr)
+			}
+			return nil, err
+		}
+	}
+
+	err = macaroonService.Close()
+	if err != nil {
+		return nil, fmt.Errorf("could not close macaroon service: %v",
+			err)
+	}
+
 	// Finally, send the new password across the UnlockPasswords channel to
 	// automatically unlock the wallet.
-	u.UnlockMsgs <- &WalletUnlockMsg{Passphrase: in.NewPassword}
+	walletUnlockMsg := &WalletUnlockMsg{
+		Passphrase:    in.NewPassword,
+		Wallet:        w,
+		StatelessInit: in.StatelessInit,
+		UnloadWallet:  loader.UnloadWallet,
+	}
+	select {
+	case u.UnlockMsgs <- walletUnlockMsg:
+		// We need to read from the channel to let the daemon continue
+		// its work and to get the admin macaroon. Once the response
+		// arrives, we directly forward it to the client.
+		orderlyReturn = true
+		select {
+		case adminMac := <-u.MacResponseChan:
+			return &lnrpc.ChangePasswordResponse{
+				AdminMacaroon: adminMac,
+			}, nil
 
-	return &lnrpc.ChangePasswordResponse{}, nil
+		case <-ctx.Done():
+			return nil, ErrUnlockTimeout
+		}
+
+	case <-ctx.Done():
+		return nil, ErrUnlockTimeout
+	}
 }
 
-// validatePassword assures the password meets all of our constraints.
-func validatePassword(password []byte) error {
+// ValidatePassword assures the password meets all of our constraints.
+func ValidatePassword(password []byte) error {
 	// Passwords should have a length of at least 8 characters.
 	if len(password) < 8 {
 		return errors.New("password must have at least 8 characters")
